fix test_error
